1. Добавил в board_hardware второй мотор 2. Добавил сохранение конфигурационного файла 3. Изменил диапазон скоростей мотора

+++ raspberryApi/motor_controller.py
@@ -12,5 +12,5 @@ class Controller:
         GPIO.setmode(GPIO.BCM)
         self._pin = motor_pin
 
-    def set_speed(self, speed: float):
+    def set_speed(self, speed: int):
         pass
